Isolate Diff and convert it to FileDiff for clarity
